refactor generic store
