+++ src/ros/nova-gui/nova-gui/src/redux/models/genericStores/GenericStoreState.ts
@@ -1,0 +1,6 @@
+
+export interface GenericStoreState<T> {
+  value: T,
+}
+
+// add any generic store models here
