+++ src/ros/nova-gui/nova-gui/src/redux/store/createGenericStore.ts
@@ -1,9 +1,5 @@
 import {createSlice} from "@reduxjs/toolkit";
 
-export interface GenericStoreState<T> {
-  value: T,
-}
-
 export const createGenericStore = <T>(name: string, initialValue: T) => {
   return createSlice({
     name: name,
